+++ indico/modules/attachments/controllers/management/event.py
@@ -14,11 +14,14 @@ from indico.modules.attachments.controllers.management.base import (AddAttachmen
                                                                     DeleteAttachmentMixin, DeleteFolderMixin,
                                                                     EditAttachmentMixin, EditFolderMixin,
                                                                     ManageAttachmentsMixin)
+from indico.modules.attachments.forms import SetUploadPermissionsForm
+from indico.modules.attachments.settings import attachments_settings
 from indico.modules.attachments.util import can_manage_attachments
 from indico.modules.attachments.views import WPEventAttachments, WPPackageEventAttachmentsManagement
-from indico.modules.events.controllers.base import RHEventBase
+from indico.modules.events.controllers.base import EditEventSettingsMixin, RHEventBase
 from indico.modules.events.management.controllers import RHManageEventBase
 from indico.modules.events.util import check_event_locked, get_object_from_args
+from indico.util.i18n import _
 from indico.web.flask.templating import get_template_module
 from indico.web.rh import RHProtected
 
@@ -97,3 +100,12 @@ class RHPackageEventAttachmentsManagement(AttachmentPackageMixin, RHManageEventB
     wp = WPPackageEventAttachmentsManagement
     management = True
     ALLOW_LOCKED = True
+
+
+class RHSetUploadPermissions(EditEventSettingsMixin, RHManageEventBase):
+    settings_proxy = attachments_settings
+    form_cls = SetUploadPermissionsForm
+    success_message = _('Upload permissions changed successfully')
+    log_module = 'Attachments'
+    log_message = 'Upload settings updated'
+    log_fields = {'managers_only': 'Managers only'}